+++ repository_template/src/.contrib-candidates/yostate/tests/test_locators.py
@@ -1,0 +1,7 @@
+from ..locators import Locator
+
+
+def test_pydantic_parse_obj():
+    Locator.parse_obj({
+        'state_class_locator': '/main-menu/',
+    })