+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Update 2023-10-09
 
 
+- Yostate: исправлен баг в методе `Locator.parse_obj`
 - Добавлено django приложение `trigger_funnel`
 - Перенастроено логирование в проекте Django, чтобы сообщения долетали до консоли
 - В Django Tg Bot Framework исправлен баг обработки событий `PrivateChatMessageEdited` и `PrivateChatCallbackQuery` - #35